Criado página de recompensas

+++ src/main/java/visnode/repository/ChallengeRepository.java
@@ -46,7 +46,7 @@ public class ChallengeRepository {
      * @return Challenge
      * @throws RepositoryException
      */
-    public Challenge get(int idChallenge) throws RepositoryException {
+    public Challenge get(long idChallenge) throws RepositoryException {
         try {
             return WebService.get().
                     get("challenge",
